Add exercise to handle errors with Result<T> instead of exceptions
An in-class programming assignment was added. The objective of the assignment is to educate students on handling errors using Result<T> instead of exceptions. The exercise starts initially with a basic user authentication service that currently throws exceptions. The task is to replace these thrown exceptions with the use of Result<T> to handle errors in a more functional way.